Added/Updated: Extra sanity checks to stop running loni job.
Case 5 for jumping to filter-vcf step if the sorted bam file and
its index file have already been moved to main output dir, if it's
a multi-key blinded ID.

Added creation of directory for multikey blinded-IDs and removed
the append of the # of the file so that headers in the sorted bam
files are consistent for the mpileup step.

Error report for grabTrio-dir.sh to help debug any issues.

**Backup**
8-17-17

*End of semester backup for any issues that may occur while not on site*